Pass STATICX_VERSION from setup.py to SCons
This avoids the SCons Python environment having to deal with importing
Python code which should only concern the setup.py environment.

Also, inline dynversion.py back into setup.py.
