+++ bootloader/SConscript
@@ -2,17 +2,9 @@ import sys
 
 Import('env')
 
-projdir = env.Dir('.').srcnode().dir
-sys.path.append(projdir)
-from dynversion import get_dynamic_version
-
-
 env.Append(
     CCFLAGS = ['-static'],
     LINKFLAGS = ['-static'],
-    CPPDEFINES = dict(
-        STATICX_VERSION = '\\"{}\\"'.format(get_dynamic_version()),
-    ),
     CPPPATH = [
         '#libtar',
         '#libxz',
